+++ src/include/type/type_util.h
@@ -23,7 +23,7 @@ class TypeUtil {
    * without this constraint, but it would likely incur a speed impact on creation of ProjectedColumns and RowViews.
    * @throw std::runtime_error if type is unknown
    */
-  static uint8_t GetTypeSize(const TypeId type_id) {
+  static uint16_t GetTypeSize(const TypeId type_id) {
     switch (type_id) {
       case TypeId::BOOLEAN:
       case TypeId::TINYINT:
